requirements.txt
Django==3.1.*
django-filter==2.3.*
djangorestframework==3.11.*
Pillow==7.2.*
psycopg2-binary=2.8.*
python-dotenv==0.14.*
sorl-thumbnail==12.6.*